requirements.txt
anyio==3.6.2
ariadne==0.17.0
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
dnspython==2.2.1
ecdsa==0.18.0
Flask==2.2.2
flask-oidc==1.4.0
graphql-core==3.2.3
h11==0.12.0
httpcore==0.13.7
httplib2==0.21.0
httpx==0.18.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.25.0
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymongo==4.3.3
pyparsing==3.0.9
python-dotenv==0.21.0
python-jose==3.3.0
python-keycloak==2.8.0
PyYAML==5.4.1
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.23.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
